+++ src/modules/Group.ts
@@ -20,25 +20,26 @@ export interface IGroup {
   lastUpdatedAt: string;
 }
 
-export interface IClassGroup {
-  details: IGroup;
-  remove(): Promise<number>;
-}
+export type IGroupUpdate =
+  | {
+      id: string;
+    }
+  | { name: string };
 
-export class Group implements IClassGroup {
+export class Group {
   private id: string;
   private saasClient: QlikSaaSClient;
   details: IGroup;
   constructor(saasClient: QlikSaaSClient, id: string, details?: IGroup) {
     if (!id) throw new Error(`group.get: "id" parameter is required`);
 
+    this.details = details ?? ({} as IGroup);
     this.id = id;
     this.saasClient = saasClient;
-    if (details) this.details = details;
   }
 
-  async init() {
-    if (!this.details) {
+  async init(arg?: { force: true }) {
+    if (Object.keys(this.details).length == 0 || arg?.force == true) {
       this.details = await this.saasClient
         .Get(`groups/${this.id}`)
         .then((res) => res.data as IGroup);
@@ -49,5 +50,25 @@ export class Group implements IClassGroup {
     return await this.saasClient
       .Delete(`groups/${this.id}`)
       .then((res) => res.status);
+  }
+
+  async update(arg: IGroupUpdate[]) {
+    if (!arg) throw new Error(`group.update: update arguments are missing`);
+
+    const data = arg.map((a) => ({
+      path: `/assignedRoles`,
+      value: a,
+      op: "replace",
+    }));
+
+    let updateStatus: number = -1;
+
+    return await this.saasClient
+      .Patch(`groups/${this.id}`, data)
+      .then((res) => {
+        updateStatus = res.status;
+        return this.init({ force: true });
+      })
+      .then(() => updateStatus);
   }
 }